Add initial test data

+++ main.py
@@ -3,6 +3,10 @@ import requests
 import os
 import json
 import uuid
+import time
+from tqdm import tqdm
+import subprocess
+import filecmp
 
 def download_audio(url, folder, filename):
     response = requests.get(url)
@@ -11,7 +15,7 @@ def download_audio(url, folder, filename):
         file_path = os.path.join(folder, filename)
         with open(file_path, 'wb') as file:
             file.write(response.content)
-        print(f"Audio file downloaded and saved as '{file_path}'")
+        # print(f"Audio file downloaded and saved as '{file_path}'")
     else:
         print("Failed to download audio file")
 
@@ -37,7 +41,7 @@ def get_audio_link(text):
     match = re.search(pattern, response.text.split("event: completed")[1])
     return match.group(1)
 
-def convert_json(input_file, output_file):
+def convert_arrow_json(input_file, output_file):
     with open(input_file, 'r') as f:
         data = json.load(f)
 
@@ -82,23 +86,28 @@ def convert_json(input_file, output_file):
 
 
 def save_text_file(file_path, content):
+    if os.path.exists(file_path):
+        return False
+    
     try:
         with open(file_path, 'w') as file:
             file.write(content)
-        print(f"Text file '{file_path}' saved successfully.")
+        # print(f"Text file '{file_path}' saved successfully.")
     except IOError:
         print(f"Error: Unable to save text file '{file_path}'.")
+
+    return True
 
 
 def create_folder_if_not_exists(folder_path):
     if not os.path.exists(folder_path):
         try:
             os.makedirs(folder_path)
-            print(f"Folder '{folder_path}' created successfully.")
+            # print(f"Folder '{folder_path}' created successfully.")
         except OSError:
             print(f"Error: Unable to create folder '{folder_path}'.")
 
-def setup_directory(name):
+def pre_process_story(name):
 
     input = f"stories/{name}/raw_input.json"
     output = f"stories/{name}/processed_input.json"
@@ -107,27 +116,67 @@ def setup_directory(name):
         print("Couldn't find raw_input.json... Try again by creating a folder and adding the arrows.app starter file")
         exit()
     
-    if not os.path.exists(output):
-        convert_json(input, output)
+    convert_arrow_json(input, output)
 
     folder_names = ["text", "audio", "gentle"]
     for folder_name in folder_names:
         folder_path = os.path.join(f'stories/{name}', folder_name)
         create_folder_if_not_exists(folder_path)
 
-def save_clip(id, text, folder):
-    save_text_file(f'stories/{folder}/text/{id}.txt', text)
-    download_audio(get_audio_link(sample), f'stories/{story_name}/audio', f'{id}.mp3')
+def save_segment(id, text, folder):
+    new_file = save_text_file(f'stories/{folder}/text/{id}.txt', text)
+    if new_file or not os.path.exists(f'stories/{story_name}/audio/{id}.mp3'):
+        download_audio(get_audio_link(text), f'stories/{story_name}/audio', f'{id}.mp3')
 
+    if new_file or not os.path.exists(f'stories/{story_name}/gentle/{id}.json'):
+        command = ['python', 'gentle/align.py', f'stories/{story_name}/audio/{id}.mp3', f'stories/{story_name}/text/{id}.txt', '-o', f'stories/{story_name}/gentle/{id}.json']
+        try:
+            subprocess.run(command, check=True)
+            # print("Alignment completed successfully.")
+        except subprocess.CalledProcessError as e:
+            print("Alignment failed with error:", e)
+
+def download_story_components(name):
+    story_json = f"stories/{name}/processed_input.json"
+    with open(story_json, 'r') as f:
+        data = json.load(f)
+        
+        print("Deleting existing extra files")
+        filenames = []
+        directories = [{
+            "path": "audio",
+            "ext": "mp3"
+        },
+        {
+            "path": "gentle",
+            "ext": "json"
+        },
+        {
+            "path": "text",
+            "ext": "txt"
+        }]
+        for node in data['nodes']:
+            filenames.append(data['nodes'][node]["id"])
+        for directory in directories:
+            for filename in os.listdir(f'stories/{name}/{directory["path"]}'):
+                if not filename.endswith(directory["ext"]) or filename.split("/")[-1].split(".")[0] not in filenames:
+                    file_path = os.path.join(f'stories/{name}/{directory["path"]}', filename)
+                    os.remove(file_path)
+                    print(f"Deleted file: {file_path}")
+
+        print("Downloading story content")
+        for node in tqdm(data['nodes']):
+            id = data['nodes'][node]["id"]
+            text = node
+            folder = name
+            save_segment(id, text, folder)
+
+def process_story(name):
+    pre_process_story(name)
+    download_story_components(name)
 
 
 # Setup
 # story_name = input("What story are we working on: ")
 story_name = "test"
-setup_directory(story_name)
-
-# Download clip
-# id = "test"
-# story_name = "test"
-# sample = "Despite bring over 3 inches in length, the tarantula is not large enough to have a measurable gravitational pull on the Sun."
-# save_clip(id, sample, story_name)
+process_story(story_name)
